+++ public/components/search_configuration/views/search_configuration_create.tsx
@@ -44,6 +44,10 @@ export const SearchConfigurationCreate: React.FC<SearchConfigurationCreateProps>
     setName,
     nameError,
     validateNameField,
+    description,
+    setDescription,
+    descriptionError,
+    validateDescriptionField,
     query,
     setQuery,
     queryError,
@@ -129,6 +133,10 @@ export const SearchConfigurationCreate: React.FC<SearchConfigurationCreateProps>
               setName={setName}
               nameError={nameError}
               validateName={validateNameField}
+              description={description}
+              setDescription={setDescription}
+              descriptionError={descriptionError}
+              validateDescription={validateDescriptionField}
               query={query}
               setQuery={setQuery}
               queryError={queryError}